vagrant/jenkins use docker and fix reference issue
Turns out the version downloaded via apt-get is too new, which does not
allow usernames with dots. The version in Docker does, although it's
older. That might be something to fix later.

The for loop that configured the VMs was causing the jk reference to
point to the last VM, so doing vagrant ssh jenkins2 was not working.
That's fixed now by removing the for loop and using separate statements.